[정렬]9월16일

+++ 9월3일-정렬/13458.cpp
@@ -5,14 +5,14 @@ using namespace std;
 
 vector <int> arr;
 
-double peo(int n, int a, int b) {
-    double count = 0;
+long peo(int n, int a, int b) {
+    long count = 0;
     for (int i = 0; i < n; i++) {
         int num = arr[i] - a;
-        
+
         int d = num / b;
-        if (num <0) count = 1;
-        else if ((num %b) == 0) count += d + 1;
+        if (num < 0) count = 1;
+        else if ((num % b) == 0) count += d + 1;
         else count += d + 2;
     }
     return count;
